+++ include/CCeilLProc.h
@@ -76,7 +76,7 @@ class ClLanguageProc {
  public:
   const std::string &getName() const { return name_; }
 
-  uint getNumArgs() const { return args_.size(); }
+  uint getNumArgs() const { return uint(args_.size()); }
 
   const ClLanguageProcArg &getArg(uint i) const {
     assert(i < args_.size());
